Extract local brand detection logic into separate class
COSDK-921
